requirements.txt
Pillow==9.5.0
PySimpleGUI==4.60.5
